Improve Code Quality (Parser)
Implemented the use of constants and removed magic strings. This will
help with the readibility and accessbility of the code in the future as
the named constants makes the code easier to understand.

Decreased deep nesting by abstracting various parts of the validation
logic into its respective methods. This helps improve the readbility and
accessbility of the code as each it is now implicitly stated what
validation is being conducted