Implement Update Command
Added parser logic for update command. update takes in a mandatory field
index, while there are also multiple optional field,
/t <tasktype> used to declare a change in task type
/d <description> used to declare a new description
/by <DateTimeLocal> used to declare a new by. (This field is required if
a user is looking to change the current task types to be a Deadline
task)
/from <DateTimeLocal> used to declare a new from. (This field is
required if a user is looking to change the current Task to an Event
task)
/to <DateTimeLocal> used to declare a new to. (This field is required if
a user is looking to change the current Task to an Event Task)
*Do note, users are required to use at least one optional field when
using update)

Added UpdateCommand, like the other command types, it handles the logic
of excecuting the task. UpdateCommand uses the given inputs above and
creates a new updatedTask, after which it replaces the selectedTask with
it, thus updating the task.

Created getter for getBy for DeadlineTask
Created getters getFrom, getTo for EventTask
Created setIndex method in TaskList which takes a task and the index to
replace it at.