Fix bugs & Add customized line & style for Update
Fixed bugs that would result in the error-label not display. Had to rest
command to null upon catching error. Added additional check for update
command to throw an error if all args are null. This is also ensures
that the error-label is used when displaying the message.

Added customized dialog lines upon completing success update